create insert method to db

+++ src/main/java/payroll/TransactionRepository.java
@@ -3,6 +3,6 @@ package payroll;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.repository.CrudRepository;
 
-public interface TransactionRepository extends CrudRepository<TransactionModel, Long> {
+public interface TransactionRepository extends CrudRepository<TransactionModel, Integer> {
 
 }
